LET THERE BE A RUNTIME!
Split out runtime stuff into a separate package
