+++ pkg/manifest/manifest_test.go
@@ -4,16 +4,124 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/deislabs/porter/pkg/context"
-
-	"github.com/deislabs/porter/pkg/config"
-
 	"github.com/deislabs/cnab-go/bundle/definition"
-	"gopkg.in/yaml.v2"
-
+	"github.com/deislabs/porter/pkg/config"
+	"github.com/deislabs/porter/pkg/context"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"gopkg.in/yaml.v2"
 )
+
+func TestLoadManifest(t *testing.T) {
+	cxt := context.NewTestContext(t)
+
+	cxt.AddTestFile("testdata/simple.porter.yaml", config.Name)
+
+	m, err := LoadManifestFrom(cxt.Context, config.Name)
+	require.NoError(t, err, "could not load manifest")
+
+	require.NotNil(t, m, "manifest was nil")
+	assert.Equal(t, []MixinDeclaration{{Name: "exec"}}, m.Mixins, "expected manifest to declare the exec mixin")
+	require.Len(t, m.Install, 1, "expected 1 install step")
+
+	installStep := m.Install[0]
+	description, _ := installStep.GetDescription()
+	assert.NotNil(t, description, "expected the install description to be populated")
+
+	mixin := installStep.GetMixinName()
+	assert.Equal(t, "exec", mixin, "incorrect install step mixin used")
+
+	require.Len(t, m.CustomActions, 1, "expected manifest to declare 1 custom action")
+	require.Contains(t, m.CustomActions, "status", "expected manifest to declare a status action")
+
+	statusStep := m.CustomActions["status"][0]
+	description, _ = statusStep.GetDescription()
+	assert.Equal(t, "Get World Status", description, "unexpected status step description")
+
+	mixin = statusStep.GetMixinName()
+	assert.Equal(t, "exec", mixin, "unexpected status step mixin")
+}
+
+func TestLoadManifestWithDependencies(t *testing.T) {
+	cxt := context.NewTestContext(t)
+
+	cxt.AddTestFile("testdata/porter.yaml", config.Name)
+	cxt.AddTestDirectory("testdata/bundles", "bundles")
+
+	m, err := LoadManifestFrom(cxt.Context, config.Name)
+	require.NoError(t, err, "could not load manifest")
+
+	assert.NotNil(t, m)
+	assert.Equal(t, []MixinDeclaration{{Name: "exec"}}, m.Mixins)
+	assert.Len(t, m.Install, 1)
+
+	installStep := m.Install[0]
+	description, _ := installStep.GetDescription()
+	assert.NotNil(t, description)
+
+	mixin := installStep.GetMixinName()
+	assert.Equal(t, "exec", mixin)
+}
+
+func TestAction_Validate_RequireMixinDeclaration(t *testing.T) {
+	cxt := context.NewTestContext(t)
+
+	cxt.AddTestFile("testdata/simple.porter.yaml", config.Name)
+
+	m, err := LoadManifestFrom(cxt.Context, config.Name)
+	require.NoError(t, err, "could not load manifest")
+
+	// Sabotage!
+	m.Mixins = []MixinDeclaration{}
+
+	err = m.Install.Validate(m)
+	assert.EqualError(t, err, "mixin (exec) was not declared")
+}
+
+func TestAction_Validate_RequireMixinData(t *testing.T) {
+	cxt := context.NewTestContext(t)
+
+	cxt.AddTestFile("testdata/simple.porter.yaml", config.Name)
+
+	m, err := LoadManifestFrom(cxt.Context, config.Name)
+	require.NoError(t, err, "could not load manifest")
+
+	// Sabotage!
+	m.Install[0].Data = nil
+
+	err = m.Install.Validate(m)
+	assert.EqualError(t, err, "no mixin specified")
+}
+
+func TestAction_Validate_RequireSingleMixinData(t *testing.T) {
+	cxt := context.NewTestContext(t)
+
+	cxt.AddTestFile("testdata/simple.porter.yaml", config.Name)
+
+	m, err := LoadManifestFrom(cxt.Context, config.Name)
+	require.NoError(t, err, "could not load manifest")
+
+	// Sabotage!
+	m.Install[0].Data["rando-mixin"] = ""
+
+	err = m.Install.Validate(m)
+	assert.EqualError(t, err, "more than one mixin specified")
+}
+
+func TestManifest_Validate_Dockerfile(t *testing.T) {
+	cxt := context.NewTestContext(t)
+
+	cxt.AddTestFile("testdata/simple.porter.yaml", config.Name)
+
+	m, err := LoadManifestFrom(cxt.Context, config.Name)
+	require.NoError(t, err, "could not load manifest")
+
+	m.Dockerfile = "Dockerfile"
+
+	err = m.Validate()
+
+	assert.EqualError(t, err, "Dockerfile template cannot be named 'Dockerfile' because that is the filename generated during porter build")
+}
 
 func TestReadManifest_URL(t *testing.T) {
 	cxt := context.NewTestContext(t)
