+++ pkg/runtime/actionInput.go
@@ -1,0 +1,36 @@
+package runtime
+
+import (
+	"github.com/deislabs/porter/pkg/manifest"
+	"gopkg.in/yaml.v2"
+)
+
+type ActionInput struct {
+	action manifest.Action
+	Steps  []*manifest.Step `yaml:"steps"`
+}
+
+// MarshalYAML marshals the step nested under the action
+// install:
+// - helm:
+//   ...
+// Solution from https://stackoverflow.com/a/42547226
+func (a *ActionInput) MarshalYAML() (interface{}, error) {
+	// encode the original
+	b, err := yaml.Marshal(a.Steps)
+	if err != nil {
+		return nil, err
+	}
+
+	// decode it back to get a map
+	var tmp interface{}
+	err = yaml.Unmarshal(b, &tmp)
+	if err != nil {
+		return nil, err
+	}
+	stepMap := tmp.([]interface{})
+	actionMap := map[string]interface{}{
+		string(a.action): stepMap,
+	}
+	return actionMap, nil
+}
