+++ pkg/runtime/actionInput_test.go
@@ -1,0 +1,30 @@
+package runtime
+
+import (
+	"testing"
+
+	"github.com/deislabs/porter/pkg/manifest"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"gopkg.in/yaml.v2"
+)
+
+func TestActionInput_MarshalYAML(t *testing.T) {
+	s := &manifest.Step{
+		Data: map[string]interface{}{
+			"exec": map[string]interface{}{
+				"command": "echo hi",
+			},
+		},
+	}
+
+	input := &ActionInput{
+		action: manifest.ActionInstall,
+		Steps:  []*manifest.Step{s},
+	}
+
+	b, err := yaml.Marshal(input)
+	require.NoError(t, err)
+	wantYaml := "install:\n- exec:\n    command: echo hi\n"
+	assert.Equal(t, wantYaml, string(b))
+}
