+++ pkg/runtime/runtime_test.go
@@ -1,0 +1,228 @@
+package runtime
+
+import (
+	"fmt"
+	"path/filepath"
+	"testing"
+
+	"github.com/deislabs/cnab-go/bundle/definition"
+	"github.com/deislabs/porter/pkg/config"
+	"github.com/deislabs/porter/pkg/context"
+	"github.com/deislabs/porter/pkg/manifest"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestPorterRuntime_Execute_readMixinOutputs(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+
+	testFiles := []string{
+		"emptyoutput",
+		"jsonoutput",
+		"multiliner",
+		"oneliner",
+	}
+
+	for _, testFile := range testFiles {
+		r.TestContext.AddTestFile(
+			fmt.Sprintf("testdata/outputs/%s.txt", testFile),
+			filepath.Join(context.MixinOutputsDir, testFile))
+	}
+
+	gotOutputs, err := r.readMixinOutputs()
+	require.NoError(t, err)
+
+	for _, testFile := range testFiles {
+		if exists, _ := r.Context.FileSystem.Exists(testFile); exists {
+			require.Fail(t, fmt.Sprintf("file %s should not exist after reading outputs", testFile))
+		}
+	}
+
+	wantOutputs := map[string]string{
+		"emptyoutput": "",
+		"jsonoutput": `{
+  "foo": true,
+  "things": [
+    123,
+    "abc",
+    false
+  ]
+}`,
+		"multiliner": `FOO
+
+BAR
+BAZ`,
+		"oneliner": "ABC",
+	}
+
+	assert.Equal(t, wantOutputs, gotOutputs)
+}
+
+func TestPorterRuntime_ApplyStepOutputsToBundle_None(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	m := &manifest.Manifest{Name: "mybun"}
+	r.RuntimeManifest = NewRuntimeManifest(r.Context, manifest.ActionInstall, m)
+
+	outputs := map[string]string{
+		"foo": "bar",
+		"123": "abc",
+	}
+
+	err := r.applyStepOutputsToBundle(outputs)
+	assert.NoError(t, err)
+}
+
+func TestPorterRuntime_ApplyStepOutputsToBundle_Some_Match(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	m := &manifest.Manifest{
+		Name: "mybun",
+		Outputs: []manifest.OutputDefinition{
+			{
+				Name: "foo",
+				Schema: definition.Schema{
+					Type: "string",
+				},
+				Sensitive: true,
+			},
+			{
+				Name: "123",
+				Schema: definition.Schema{
+					Type: "string",
+				},
+				Sensitive: false,
+			},
+		},
+	}
+	r.RuntimeManifest = NewRuntimeManifest(r.Context, manifest.ActionInstall, m)
+
+	outputs := map[string]string{
+		"foo": "bar",
+		"123": "abc",
+	}
+
+	err := r.applyStepOutputsToBundle(outputs)
+	assert.NoError(t, err)
+
+	want := map[string]string{
+		"foo": "bar",
+		"123": "abc",
+	}
+
+	for _, outputName := range []string{"foo", "123"} {
+		bytes, err := r.FileSystem.ReadFile(filepath.Join(config.BundleOutputsDir, outputName))
+		assert.NoError(t, err)
+
+		assert.Equal(t, want[outputName], string(bytes))
+	}
+}
+
+func TestPorterRuntime_ApplyStepOutputsToBundle_Some_NoMatch(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	m := &manifest.Manifest{
+		Name: "mybun",
+		Outputs: []manifest.OutputDefinition{
+			{
+				Name: "bar",
+			},
+			{
+				Name: "456",
+			},
+		},
+	}
+	r.RuntimeManifest = NewRuntimeManifest(r.Context, manifest.ActionInstall, m)
+
+	outputs := map[string]string{
+		"foo": "bar",
+		"123": "abc",
+	}
+
+	err := r.applyStepOutputsToBundle(outputs)
+	assert.NoError(t, err)
+
+	// No outputs declared in the manifest match those in outputs,
+	// so no output file is expected to be written
+	for _, output := range []string{"foo", "bar", "123", "456"} {
+		_, err := r.FileSystem.ReadFile(filepath.Join(config.BundleOutputsDir, output))
+		assert.Error(t, err)
+	}
+}
+
+func TestPorterRuntime_ApplyStepOutputsToBundle_ApplyTo_True(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	m := &manifest.Manifest{
+		Name: "mybun",
+		Outputs: []manifest.OutputDefinition{
+			{
+				Name: "foo",
+				ApplyTo: []string{
+					"upgrade",
+				},
+			},
+			{
+				Name: "123",
+				ApplyTo: []string{
+					"install",
+				},
+				Schema: definition.Schema{
+					Type: "string",
+				},
+				Sensitive: false,
+			},
+		},
+	}
+	r.RuntimeManifest = NewRuntimeManifest(r.Context, manifest.ActionInstall, m)
+
+	outputs := map[string]string{
+		"foo": "bar",
+		"123": "abc",
+	}
+
+	err := r.applyStepOutputsToBundle(outputs)
+	assert.NoError(t, err)
+
+	// foo output should not exist (applyTo doesn't match)
+	_, err = r.FileSystem.ReadFile(filepath.Join(config.BundleOutputsDir, "foo"))
+	assert.Error(t, err)
+
+	// 123 output should exist (applyTo matches)
+	bytes, err := r.FileSystem.ReadFile(filepath.Join(config.BundleOutputsDir, "123"))
+	assert.NoError(t, err)
+
+	want := "abc"
+	assert.Equal(t, want, string(bytes))
+}
+
+func TestPorterRuntime_LoadImageMappingFilesNoRelo(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	r.TestContext.AddTestFile("testdata/bundle-images.json", "/cnab/bundle.json")
+	bun, reloMap, err := r.getImageMappingFiles()
+	assert.NoError(t, err)
+	assert.Empty(t, reloMap)
+	assert.Equal(t, "mysql", bun.Name)
+}
+
+func TestLoadImageMappingFilesNoBundle(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	r.TestContext.AddTestFile("testdata/relocation-mapping.json", "/cnab/app/relocation-mapping.json")
+	_, _, err := r.getImageMappingFiles()
+	assert.EqualError(t, err, "couldn't read runtime bundle.json: open /cnab/bundle.json: file does not exist")
+}
+
+func TestLoadImageMappingFilesBadBundle(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	r.TestContext.AddTestFile("../porter/testdata/porter.yaml", "/cnab/bundle.json")
+	r.TestContext.AddTestFile("testdata/relocation-mapping.json", "/cnab/app/relocation-mapping.json")
+	_, _, err := r.getImageMappingFiles()
+	assert.EqualError(t, err, "couldn't load runtime bundle.json: invalid character 'a' in literal null (expecting 'u')")
+}
+
+func TestLoadImageMappingFilesGoodFiles(t *testing.T) {
+	r := NewTestPorterRuntime(t)
+	r.TestContext.AddTestFile("testdata/bundle-images.json", "/cnab/bundle.json")
+	r.TestContext.AddTestFile("testdata/relocation-mapping.json", "/cnab/app/relocation-mapping.json")
+
+	bun, reloMap, err := r.getImageMappingFiles()
+	assert.NoError(t, err)
+	assert.NotEmpty(t, reloMap)
+	assert.Equal(t, "mysql", bun.Name)
+}
