+++ pkg/runtime/testdata/dep-metadata-substitution.yaml
