+++ pkg/runtime/testdata/metadata-substitution.yaml
