+++ pkg/runtime/testdata/outputs/oneliner.txt
