+++ pkg/runtime/testdata/relocation-mapping.json
